refactor(server): - add utils for getting coordinates and converting xml - uninstall xml2js
