+++ backend/package.json
@@ -18,7 +18,6 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "helmet": "^6.0.1",
-    "xml2js": "^0.4.23"
+    "helmet": "^6.0.1"
   }
 }
