+++ backend/src/routes/routes.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const { getWeatherData } = require("../controllers/app.controllers");
+const { getWeatherData } = require("../controllers/controllers");
 
 const Router = express.Router();
